Add a project to build the reader \ writer with internal visibility
