+++ src/MessagePack/Nil.cs
@@ -3,7 +3,12 @@ using System.Buffers;
 
 namespace MessagePack
 {
-    public struct Nil : IEquatable<Nil>
+#if MESSAGEPACK_INTERNAL
+    internal
+#else
+    public
+#endif
+    struct Nil : IEquatable<Nil>
     {
         public static readonly Nil Default = new Nil();
 
@@ -25,52 +30,6 @@ namespace MessagePack
         public override string ToString()
         {
             return "()";
-        }
-    }
-}
-
-namespace MessagePack.Formatters
-{
-    public class NilFormatter : IMessagePackFormatter<Nil>
-    {
-        public static readonly IMessagePackFormatter<Nil> Instance = new NilFormatter();
-
-        NilFormatter()
-        {
-
-        }
-
-        public void Serialize(ref MessagePackWriter writer, Nil value, IFormatterResolver typeResolver)
-           
-        {
-            writer.WriteNil();
-        }
-
-        public Nil Deserialize(ref MessagePackReader reader, IFormatterResolver typeResolver)
-        {
-            return reader.ReadNil();
-        }
-    }
-
-    // NullableNil is same as Nil.
-    public class NullableNilFormatter : IMessagePackFormatter<Nil?>
-    {
-        public static readonly IMessagePackFormatter<Nil?> Instance = new NullableNilFormatter();
-
-        NullableNilFormatter()
-        {
-
-        }
-
-        public void Serialize(ref MessagePackWriter writer, Nil? value, IFormatterResolver typeResolver)
-           
-        {
-            writer.WriteNil();
-        }
-
-        public Nil? Deserialize(ref MessagePackReader reader, IFormatterResolver typeResolver)
-        {
-            return reader.ReadNil();
         }
     }
 }
